requirements.txt
Django>=5.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
python-dotenv>=1.0.1
requests>=2.32.0
googlemaps>=4.10.0
psycopg2-binary>=2.9.9  # PostgreSQL driver (if using PostgreSQL; skip if using SQLite)
Pillow>=10.2.0          # For handling images if needed later
django-cors-headers
langchain_ollama